Fixed mutator mutation order.
Assuming the mutator module is correct*, mutations on get are applied in
the reverse order from put.

A mutator module is "correct" if the get is the reverse function of a put.
All the data a mutator needs to reverse a given mutation must be put into
the object to be able to do this.